Uploaded Element Pre-Delete
This script prompts the user to select an element in Revit. It will then run (but not commit) a transaction to delete this element. The transaction is rolled-back - but we are able to get the ElementIds of all of the elements which WOULD have been deleted had this transaction been committed. We can use the ElementIds to see which elements are dependent on the user-specified element.