Create rdsCsvExporter.py
for large projects with lots of room data it is useful to analyse in
a business inteligence or data visualisation tool such as ms power bi.
This data could be simply exported from revit as a room schedule,
however for a single power bi template to work for many projects across
a practice the room data needs to be formatted in a consistent way that
would be subject to user error. This script exports a csv file of room
data. The user can define what parameters to use which can easily be
changed in dynamo player.