Update roomSvgExporter.py
Hello,

Thanks for publishing roomSvgExporter. It has been very helpful.

However, during development I found that when a model contains Rooms and Spaces they are both exported. This results in doubling up the SVG output when Room and Space boundaries match which was unexpected.

I don't understand the reason why but it seems that the Revit API treats Rooms and Spaces the same when it comes to built in parameters like area. So filtering by area isn't enough to differentiate Rooms from Spaces. As a result I added a Category filter to export desired SVG data: Room or Space but not both. Please review the category filter modification in particular. It seems to work OK but I have very little experience writing Dynamo Python scripts. 

Finally, an output directory input was added to improve control over the output destination. 

Regards,
Jake Staub